add record class combined by vector and label

+++ src/freakDatum.cpp
@@ -2,9 +2,10 @@
 #include <iostream>
 #include <algorithm>
 
-void freak::printD (const FreakVectorF& t){
+void freak::printD (const FreakVectorF& t)
+{
     std::cout<<"size : "<<t.size()<<std::endl;
-    for(int i=0;i<t.size();++i)  {
+    for(size_t i=0; i<t.size(); ++i)  {
         std::cout<<t.getByIndex(i)<<std::endl;
     }
 }
